ref #Issue-changes eigene Änderungen

+++ org.dieschnittstelle.jee.esa.ejb.ejbmodule.crm/src/main/java/org/dieschnittstelle/jee/esa/ejb/ejbmodule/crm/crud/TouchpointCRUDLocal.java
@@ -17,6 +17,6 @@ public interface TouchpointCRUDLocal {
 	
 	public AbstractTouchpoint updateTouchpoint(AbstractTouchpoint Touchpoint);
 		
-	public boolean deleteTouchpoint(int id);
+	public boolean deleteTouchpoint(long id);
 
 }
