+++ org.dieschnittstelle.jee.esa.ejb.ejbmodule.crm/src/main/java/org/dieschnittstelle/jee/esa/ejb/ejbmodule/crm/crud/TouchpointCRUDRemote.java
@@ -17,6 +17,6 @@ public interface TouchpointCRUDRemote {
 	
 	public AbstractTouchpoint updateTouchpoint(AbstractTouchpoint Touchpoint);
 		
-	public boolean deleteTouchpoint(int id);
+	public boolean deleteTouchpoint(long id);
 
 }
